% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetR-package.R
\docType{package}
\name{vegetR-package}
\alias{vegetR}
\alias{vegetR-package}
\title{vegetR: Managing Data from Vegetation Inventories}
\description{
This package provides tools for reading, converting, processing, and saving data from quantitative vegetation inventories, such as inventory methods, site descriptions, species records, measurements and lists.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LimaRAF/vegetR}
  \item Report bugs at \url{https://github.com/LimaRAF/vegetR/issues}
}

}
\author{
\strong{Maintainer}: Renato Lima \email{raflima@usp.br} (\href{https://orcid.org/0000-0002-1048-0138}{ORCID}) [copyright holder]

}
\keyword{internal}
